Improved EnemySpawning even further
-Fixed the calculation so 2 enemies now start spawning at once starting at Level 4
-Varied enemy speeds even more by making them floats and choosing a random decimal value in between
-Passed SerializationException into the catch statement when loading data
-Made levels slightly longer over time
-Removed HUDBottomBounds from Player.cs

+++ Defend Your Castle/Defend Your Castle/Enemies/ArmoredEnemy.cs
@@ -10,9 +10,9 @@ namespace Defend_Your_Castle
     //An armored enemy
     public sealed class ArmoredEnemy : Enemy
     {
-        public ArmoredEnemy(Level level, float Y, int speedadd, int costume)
+        public ArmoredEnemy(Level level, float Y, float speedadd, int costume)
         {
-            MoveSpeed = new Vector2(1 + speedadd, 0);
+            MoveSpeed = new Vector2(1f + speedadd, 0);
 
             WeaponWeakness = (int)Player.WeaponTypes.Warhammer;
