+++ Defend Your Castle/Defend Your Castle/Enemies/SpearEnemy.cs
@@ -13,9 +13,9 @@ namespace Defend_Your_Castle
         //The amount of damage the projectile deals
         protected int ProjectileDamage;
 
-        public SpearEnemy(Level level, float Y, int speedadd, int costume)
+        public SpearEnemy(Level level, float Y, float speedadd, int costume)
         {
-            MoveSpeed = new Vector2(1 + speedadd, 0);
+            MoveSpeed = new Vector2(1f + speedadd, 0);
             Range = 150;
 
             ProjectileDamage = 25;
